Made a more readable pipeline for modelling.
Results are included in the NoAIresults folder, there are 3 versions, the code that produced the graphs and output is included because github is a lie from illuminati. Other stuff that needs to be done is included as TODO at the beginning of the code. Will do most of them tomorrow